added recycler view to show list of countries

+++ app/src/main/java/com/dynamicdudes/covid_19/data/CountryInfo.kt
@@ -15,7 +15,7 @@ data class CountryInfo(
     @SerializedName("iso3")
     val iso3: String,
     @SerializedName("lat")
-    val lat: Int,
+    val lat: Double,
     @SerializedName("long")
-    val long: Int
+    val long: Double
 )
